Add the core logic to pull the button states from the SNES pad
Adding the simple function to read from the pad's shift register.
Here is how it works:

* Send a pulse on the latch pin, the register stores the buttons' states in the register

* Read the value on the data pin (and store it)

* Send a pulse to the clock pin, the register shifts out the next bit of data

* Repeat steps 2 and 3 until all 12 states have been pulled from the register

To make further processing easier, received bits are returned inverted: When a button is pressed during the latch pulse, it sets the corresponding bit to 0 (logic low).
Also, only 12 of the 16 bits in `uint16_t state` are used, so the 4 most significant bits will always be returned as 1, thus they must be ignored.
A future commit will include constants to help parsing the returned data, some examples will also be added.