Define constants for parsing state-values
Using the BTN_ constants, you can easily parse the received value for the button states.
Just AND-ing the state value with a constant results in a TRUE result,
if the button for that specific constant was pushed.
This way, you can easily check for single buttons to be pressed.
`getButtons() & BTN_A` is true, if the A-Button was pushed.
`!(~getButtons() & (BTN_A | BTN_B))` is true, if BOTH A and B were pushed.
The next commit will provide an option to return non-inverted output.
That's easier to check against multiple button presses.