Add capability to return the non-inverted button states
In some situations, such as checking for many simultaneous button presses,
it could be helpful to work with the "raw" output from the gamepad.
Pass false as an argument to `uint16_t getButtons(boolean return_inverted)`
to get the non-inverted output from the controller.
You don't have to pass any arguments to getButtons, the default is true
(return inverted output).